+++ analyzer/dataset_explore/vent_explore.py
@@ -9,14 +9,14 @@ import os
 from os.path import join as osjoin
 import pandas as pd
 import yaml
-from datasets.derived_ards_dataset import MIMICIV_ARDS_Dataset
-
+from datasets.derived_vent_dataset import MIMICIV_Vent_Dataset
+from scipy.signal import convolve2d
 
 class VentFeatureExplorer:
     def __init__(self, params:dict, container:DataContainer) -> None:
         self.params = params
         self.container = container
-        self.dataset = MIMICIV_ARDS_Dataset()
+        self.dataset = MIMICIV_Vent_Dataset()
         self.dataset.load_version(params['dataset_version'])
         self.gbl_conf = container._conf
         self.data = self.dataset.data
@@ -32,6 +32,11 @@ class VentFeatureExplorer:
             self.correlation(out_dir, self.params['correlation']['target'])
         if self.params['miss_mat']:
             self.miss_mat(out_dir)
+        if self.params['vent_statistics']:
+            self.vent_statistics(out_dir)
+        if self.params['vent_sample']['enabled']:
+            self.vent_sample(out_dir)
+        
     
     def correlation(self, out_dir, target_id_or_label):
         # plot correlation matrix
@@ -50,7 +55,7 @@ class VentFeatureExplorer:
     
     def miss_mat(self, out_dir):
         '''计算行列缺失分布并输出'''
-        na_table = np.ones((len(self.dataset.subjects), len(self.dataset._dynamic_keys)), dtype=bool)
+        na_table = np.ones((len(self.dataset.subjects), len(self.dataset._dynamic_keys)), dtype=bool) # True=miss
         for r_id, s_id in enumerate(self.dataset.subjects):
             for adm in self.dataset.subjects[s_id].admissions:
                 # TODO 替换dynamic keys到total keys
@@ -58,7 +63,7 @@ class VentFeatureExplorer:
                 for c_id, key in enumerate(self.dataset._dynamic_keys):
                     if key in adm_key:
                         na_table[r_id, c_id] = False
-        # 行缺失
+        
         row_nas = na_table.mean(axis=1)
         col_nas = na_table.mean(axis=0)
         tools.plot_single_dist(row_nas, f"Row miss rate", os.path.join(out_dir, "row_miss_rate.png"), discrete=False, adapt=True)
@@ -66,3 +71,72 @@ class VentFeatureExplorer:
         # save raw/col miss rate to file
         tools.save_pkl(row_nas, os.path.join(out_dir, "row_missrate.pkl"))
         tools.save_pkl(col_nas, os.path.join(out_dir, "col_missrate.pkl"))
+
+        # plot matrix
+        row_idx = sorted(list(range(row_nas.shape[0])), key=lambda x:row_nas[x])
+        col_idx = sorted(list(range(col_nas.shape[0])), key=lambda x:col_nas[x])
+        na_table = na_table[row_idx, :][:, col_idx] # (n_subjects, n_feature)
+        # apply conv to get density
+        conv_kernel = np.ones((3,3)) / 9
+        na_table = np.clip(convolve2d(na_table, conv_kernel, boundary='symm'), 0, 1.0)
+        tools.plot_density_matrix(1.0-na_table, 'Missing distribution for subjects and features [miss=white]', xlabel='features', ylabel='subjects',
+                               aspect='auto', save_path=os.path.join(out_dir, "miss_mat.png"))
+
+    def vent_statistics(self, out_dir):
+        result = {'no_vent':0, 'non-invasive_vent':0, 'invasive_vent':0}
+        non_invasive_vent_times = []
+        invasive_vent_times = []
+        for s_id in self.dataset.subjects:
+            subject = self.dataset.subjects[s_id]
+            if 'ventilation_num' not in subject.static_data:
+                result['no_vent'] += 1
+                continue
+            vent_num = subject.static_data['ventilation_num']
+            max_vent = int(np.max(vent_num[:, 0]))
+            if max_vent == 0:
+                result['no_vent'] += 1
+            elif max_vent == 1:
+                result['non-invasive_vent'] += 1
+                for idx in range(vent_num.shape[0]):
+                    if int(vent_num[idx, 0]) == 1:
+                        non_invasive_vent_times.append(vent_num[idx, 1])
+                        break
+            elif max_vent == 2:
+                result['invasive_vent'] += 1
+                for idx in range(vent_num.shape[0]):
+                    if int(vent_num[idx, 0]) == 2:
+                        invasive_vent_times.append(vent_num[idx, 1])
+                        break
+            else:
+                assert(0)
+        logger.info(f'All subjects: {len(self.dataset.subjects)}')
+        logger.info(f'Vent status (max vent for each sequence): {result}')
+        # plot distribution of first ventilation time
+        tools.plot_single_dist(np.asanyarray(non_invasive_vent_times), 'non-invasive first ventilation time', osjoin(out_dir, 'non_invasive_dist.png'), bins=72)
+        tools.plot_single_dist(np.asanyarray(invasive_vent_times), 'invasive first ventilation time', osjoin(out_dir, 'invasive_dist.png'), bins=72)
+
+    def vent_sample(self, out_dir):
+        n_plot = self.params['vent_sample']['n_plot']
+        n_idx = 1
+        plt.figure(figsize = (6, n_plot*2))
+        
+        for s_id in self.dataset.subjects:
+            subject = self.dataset.subjects[s_id]
+            if 'ventilation_num' not in subject.static_data:
+                continue
+            vent_status = int(np.max(subject.static_data['ventilation_num'][:, 0]))
+            if vent_status != 2:
+                continue
+            else:
+                plt.subplot(n_plot, 1, n_idx)
+                vent_status = subject.static_data['ventilation_num'][:, 0]
+                vent_start = subject.static_data['ventilation_start'][:, 0]
+                vent_end = subject.static_data['ventilation_end'][:, 0]
+                for idx in range(vent_status.shape[0]):
+                    plt.plot([vent_start[idx], vent_end[idx]], [vent_status[idx], vent_status[idx]], '-o')
+                n_idx += 1
+                if n_idx > n_plot:
+                    break
+        plt.savefig(osjoin(out_dir, 'vent_sample.png'))
+        plt.close()
+