+++ datasets/derived_vent_dataset.py
@@ -32,6 +32,7 @@ def load_ventilation_table(ventilation_csv_path:str, icu_stays_csv_path:str) -> 
             vent_dict[subject_id] = [item]
         else:
             vent_dict[subject_id].append(item)
+    logger.info(f'Ventilation subjects: {len(vent_dict)}')
     return vent_dict
 
 class MIMICIV_Vent_Dataset(MIMICIV_Core):
@@ -65,11 +66,17 @@ class MIMICIV_Vent_Dataset(MIMICIV_Core):
             for vent_dict in extra_data[subject_id]:
                 vent_start = ymdhms_convertor(vent_dict['starttime'])
                 vent_end = ymdhms_convertor(vent_dict['endtime'])
+                subject.append_static(vent_start, 'ventilation_start', vent_start)
+                subject.append_static(vent_start, 'ventilation_end', vent_end) # the time of 'end time' is just a key for search
                 if isinstance(vent_dict['ventilation_status'], str):
-                    subject.append_static(vent_start, 'ventilation_start', vent_start)
-                    subject.append_static(vent_start, 'ventilation_end', vent_end) # the time of 'end time' is just a key for search
-                    subject.append_static(vent_start, 'ventilation_num', self._loc_conf['ventilation_to_numeric'][vent_dict['ventilation_status']]) # to be numeric
-                
+                    subject.append_static(vent_start, 'ventilation_num', self._loc_conf['ventilation_to_numeric'][vent_dict['ventilation_status']])
+                else: # null or nan
+                    subject.append_static(vent_start, 'ventilation_num', 0)
+        else:
+            subject.append_static(0, 'ventilation_start', -1)
+            subject.append_static(0, 'ventilation_end', -1)
+            subject.append_static(0, 'ventilation_num', 0)
+
         return subject
     
     def on_extract_admission(self, subject:Subject, source:str, row:namedtuple):
@@ -101,7 +108,7 @@ class MIMICIV_Vent_Dataset(MIMICIV_Core):
                 if adm is not None:
                     discretizer = self._loc_conf['category_to_numeric']
                     careunit = discretizer['careunit'][row.careunit] if row.careunit in discretizer['careunit'] else discretizer['careunit']['Default']
-                    adm.append_dynamic('careunit', ymdhms_converter(row.intime), careunit)
+                    subject.append_static(ymdhms_converter(row.intime), 'careunit', careunit) # it is not exactly in [admittime, dischtime]
         else:
             assert(0)
     
@@ -182,30 +189,21 @@ class MIMICIV_Vent_Dataset(MIMICIV_Core):
         return invalid_count
 
     def on_select_admissions(self, rule:dict, subjects:dict[int, Subject]):
+        invalid_record = {'duration_positive':0, 'duration_limit':0, 'empty':0}
         for s_id in subjects:
             if not subjects[s_id].empty():
                 retain_adms = []
                 for idx, adm in enumerate(subjects[s_id].admissions):
-                    flag = 1
-                    # 检查target
-                    if flag != 0  and 'target_id' in rule and len(rule['target_id']) > 0:
-                        for target_id in rule['target_id']:
-                            if target_id not in adm.keys():
-                                flag = 0
-                    if flag != 0:
-                        start_time, end_time = None, None
-                        for id in rule['target_id']:
-                            start_time = max(adm[id][0,1], start_time) if start_time is not None else adm[id][0,1]
-                            end_time = min(adm[id][-1,1], end_time) if end_time is not None else adm[id][-1,1]
-                        dur = end_time - start_time
-                        if dur <= 0:
+                    dur = adm.dischtime - adm.admittime
+                    if dur <= 0:
+                        invalid_record['duration_positive'] += 1
+                        continue
+                    if 'duration_minmax' in rule:
+                        dur_min, dur_max = rule['duration_minmax']
+                        if not (dur > dur_min and dur < dur_max):
+                            invalid_record['duration_limit'] += 1
                             continue
-                        if 'duration_minmax' in rule:
-                            dur_min, dur_max = rule['duration_minmax']
-                            if not (dur > dur_min and dur < dur_max):
-                                continue
-                    if flag != 0:
-                        retain_adms.append(idx)
+                    retain_adms.append(idx)
                 subjects[s_id].admissions = [subjects[s_id].admissions[idx] for idx in retain_adms]
        
         pop_list = []
@@ -213,16 +211,21 @@ class MIMICIV_Vent_Dataset(MIMICIV_Core):
             subjects[s_id].del_empty_admission()
             if subjects[s_id].empty():
                 pop_list.append(s_id)
-                
+        invalid_record['empty'] += len(pop_list)
         for s_id in pop_list:
             subjects.pop(s_id)
         
+        # logger.info(f'invalid admissions without target id: {invalid_record["target"]}')
+        logger.info(f'invalid admissions without positive duration: {invalid_record["duration_positive"]}')
+        logger.info(f'invalid admissions exceed duration limitation: {invalid_record["duration_limit"]}')
         logger.info(f'remove_pass1: Deleted {len(pop_list)}/{len(pop_list)+len(subjects)} subjects')
+
         return subjects
 
     def on_remove_missing_data(self, rule:dict, subjects: dict[int, Subject]) -> dict[int, Subject]:
+        whitelist_subjects = {s_id for s_id, s in subjects.items() if 'ventilation_num' in s.static_data and np.max(s.static_data['ventilation_num'][:, 0]) > 0}
         n_iter = 0
-        while n_iter < len(rule['max_col_missrate']) or (len(post_dynamic_keys)+len(post_static_keys) != len(col_missrate)) or (len(pop_subject_ids) > 0):
+        while n_iter < len(rule['max_col_missrate']):
             # step1: create column missrate dict
             prior_static_keys = [k for s in subjects.values() for k in s.static_data.keys()]
             prior_dynamic_keys = [k for s in subjects.values() if not s.empty() for k in choice(s.admissions).keys()] # random smapling strategy
@@ -247,7 +250,8 @@ class MIMICIV_Vent_Dataset(MIMICIV_Core):
             # step4: remove invalid subject
             pop_subject_ids = set([s_id for s_id in subjects if subject_missrate[s_id] > rule['max_subject_missrate'][min(n_iter, len(rule['max_subject_missrate'])-1)] or subjects[s_id].empty()])
             for s_id in pop_subject_ids:
-                subjects.pop(s_id)
+                if subjects[s_id].empty() or (s_id not in whitelist_subjects): # do not pop non-empty subjects in white list
+                    subjects.pop(s_id)
             # step5: calculate removed subjects/columns
             logger.info(f'remove_pass2: iter[{n_iter}] Retain {len(self._subjects)}/{len(pop_subject_ids)+len(self._subjects)} subjects')
             logger.info(f'remove_pass2: iter[{n_iter}] Retain {len(post_dynamic_keys)+len(post_static_keys)}/{len(col_missrate)} keys in selected admission')
@@ -257,10 +261,15 @@ class MIMICIV_Vent_Dataset(MIMICIV_Core):
 
     def on_build_table(self, subject:Subject, key, value, t_start):
         admittime = subject.admissions[0].admittime
+        dischtime = subject.admissions[0].dischtime
         if key in ['ventilation_start', 'ventilation_end']:
-            return value - t_start # adjust time
-        if key == 'dod':
-            if abs(value+1.0)<1e-3:
+            if value > 0:
+                vent_time = value - (t_start + admittime) # adjust time
+            else:
+                return 0 if key == 'ventilation_start' else dischtime - admittime # for no ventilation patients
+            return vent_time
+        elif key == 'dod':
+            if abs(value+1.0) < 1e-3:
                 return -1
             else:
                 delta_year = np.floor(value / (24*365) - ((t_start+admittime) / (24*365))) # 经过多少年死亡, 两个都是timestamp，不需要加上1970
@@ -269,6 +278,8 @@ class MIMICIV_Vent_Dataset(MIMICIV_Core):
         elif key == 'age':
             age = (admittime + t_start) // (24*365) + 1970 - subject.birth_year # admittime从1970年开始计时
             return age
+        elif key == 'ventilation_num':
+            return value
         else:
             return value
     
@@ -286,7 +297,8 @@ class MIMICIV_Vent_Dataset(MIMICIV_Core):
             vent_num = table[vent_num_idx, :]
             vent_end = table[vent_end_idx, :]
             ticks = np.linspace(0, table.shape[1]*self._loc_conf['generate_table']['delta_t_hour'], table.shape[1])
-            addi_table[0, :] = vent_num * (vent_end <= ticks) # we assume that in no ventilation for a gap between last ventilation end and next ventilation start
+            addi_table[0, :] = vent_num * (vent_end >= ticks) # we assume that in no ventilation for a gap between last ventilation end and next ventilation start
+            addi_table[addi_table < 0] = 0 # impute -1
             tables[t_idx] = np.concatenate([table, addi_table], axis=0)
             norm_data.append(addi_table)
         # update norm dict for additional features
@@ -305,90 +317,3 @@ class MIMICIV_Vent_Dataset(MIMICIV_Core):
             'static_keys': static_keys,
             'dynamic_keys': dynamic_keys
         }
-    
-    def make_report(self, version_name, params:dict):
-        '''进行数据集的信息统计'''
-        # switch version
-        self.load_version(version_name)
-        self.mode('all')
-        out_path = os.path.join(self._paths['out_dir'], f'dataset_report_{version_name}.txt')
-        dist_dir = os.path.join(self._paths['out_dir'], 'report_dist')
-        dir_names = ['points', 'duration', 'frequency', 'value', 'from_sepsis', 'static_value']
-        tools.reinit_dir(dist_dir, build=True)
-        for name in dir_names:
-            os.makedirs(os.path.join(dist_dir, name))
-        logger.info('MIMIC-IV-ARDS: generating dataset report')
-        write_lines = []
-        if params['basic']:
-            # basic statistics
-            write_lines.append('='*10 + 'basic' + '='*10)
-            write_lines.append(f'Version: {version_name}')
-            write_lines.append(f'Static keys: {len(self.static_keys)}')
-            write_lines.append(f'Dynamic keys: {len(self.dynamic_keys)}')
-            write_lines.append(f'Subjects:{len(self)}')
-            write_lines.append(f'Static feature: {[self.fea_label(id) for id in self.static_keys]}')
-            write_lines.append(f'Dynamic feature: {[self.fea_label(id) for id in self.dynamic_keys]}')
-        if params['dynamic_dist']:
-            # dynamic feature explore
-            for id in tqdm(self.dynamic_keys, 'plot dynamic dist'):
-                fea_name = self.fea_label(id)
-                save_name = tools.remove_slash(str(fea_name))
-                write_lines.append('='*10 + f'{fea_name}({id})' + '='*10)
-                arr_points = []
-                arr_duration = []
-                arr_frequency = []
-                arr_avg_value = []
-                arr_from_sepsis_time = []
-                for s in self._subjects.values():
-                    for adm in s.admissions:
-                        if id in adm.keys():
-                            dur = adm[id][-1,1] - adm[id][0,1]
-                            sepsis_time = s.latest_static('sepsis_time', adm[id][0, 1])
-                            t_sep = adm[id][0, 1] + adm.admittime - sepsis_time
-                            arr_from_sepsis_time.append(t_sep)
-                            arr_points.append(adm[id].shape[0])
-                            arr_duration.append(dur)
-                            if dur > 1e-3: # TODO 只有一个点无法计算
-                                arr_frequency.append(arr_points[-1] / arr_duration[-1])
-                            else:
-                                arr_frequency.append(0)
-                            arr_avg_value.append(adm[id][:,0].mean())
-                arr_points, arr_duration, arr_frequency, arr_avg_value = \
-                    np.asarray(arr_points), np.asarray(arr_duration), np.asarray(arr_frequency), np.asarray(arr_avg_value)
-                arr_from_sepsis_time = np.asarray(arr_from_sepsis_time)
-                if np.size(arr_points) != 0:
-                    write_lines.append(f'average points per admission: {arr_points.mean():.3f}')
-                if np.size(arr_duration) != 0:
-                    write_lines.append(f'average duration(hour) per admission: {arr_duration.mean():.3f}')
-                if np.size(arr_frequency) != 0:
-                    write_lines.append(f'average frequency(point/hour) per admission: {arr_frequency.mean():.3f}')
-                if np.size(arr_avg_value) != 0:
-                    write_lines.append(f'average avg value per admission: {arr_avg_value.mean():.3f}')
-                # plot distribution
-                titles = ['points', 'duration', 'frequency', 'value', 'from_sepsis']
-                arrs = [arr_points, arr_duration, arr_frequency, arr_avg_value, arr_from_sepsis_time]
-                for title, arr in zip(titles, arrs):
-                    if np.size(arr) != 0:
-                        tools.plot_single_dist(
-                            data=arr, data_name=f'{title}: {fea_name}', 
-                            save_path=os.path.join(dist_dir, title, save_name + '.png'), discrete=False, adapt=True, bins=50)
-        if params['static_dist']:
-            # static feature explore
-            for id in tqdm(self.static_keys, 'generate static feature report'):
-                fea_name = self.fea_label(id)
-                save_name = tools.remove_slash(str(fea_name))
-                write_lines.append('='*10 + f'{fea_name}({id})' + '='*10)
-                idx = self.idx_dict[str(id)]
-                static_data = self.data[:, idx, 0]
-                write_lines.append(f'mean: {static_data.mean():.3f}')
-                write_lines.append(f'std: {static_data.std():.3f}')
-                write_lines.append(f'max: {np.max(static_data):.3f}')
-                write_lines.append(f'min: {np.min(static_data):.3f}')
-                tools.plot_single_dist(
-                    data=static_data, data_name=f'{fea_name}', 
-                    save_path=os.path.join(dist_dir, 'static_value', save_name + '.png'), discrete=False, adapt=True, bins=50)
-        # write report
-        with open(out_path, 'w', encoding='utf-8') as fp:
-            for line in write_lines:
-                fp.write(line + '\n')
-        logger.info(f'Report generated at {out_path}')